Added Screen sealed class for navigation

+++ app/src/main/java/de/alekseipopov/fooddiary/data/DayRecordRepository.kt
@@ -7,7 +7,7 @@ interface DayRecordRepository {
     suspend fun getDayRecords(): Flow<List<Day>>
     suspend fun getDay(id: Int): Flow<Day>
     suspend fun getReport(startDate: Long, endDate: Long): Flow<List<Day>>
-    suspend fun createNewDay(date: Long): Long
+    suspend fun createNewDay(date: Long): Int
     suspend fun updateDay(day: Day)
     suspend fun deleteDay(day: Day)
 }
