+++ app/src/main/java/de/alekseipopov/fooddiary/ui/overview/model/OverviewUiEvents.kt
@@ -5,6 +5,6 @@ sealed class OverviewUiEvents() {
     class HideReportDatePickerDialog() : OverviewUiEvents()
     class ShowNewEntryDialog() : OverviewUiEvents()
     class HideNewEntryDialog() : OverviewUiEvents()
-    data class ShowNewDay(val id: Long): OverviewUiEvents()
+    data class ShowNewDay(val id: Int): OverviewUiEvents()
 
 }
